pyproject.toml
[project]
name = "chatbot-planetas"
version = "0.1.0"
description = "Um mini chatbot com interface web que responde perguntas sobre planetas do Sistema Solar."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "google-generativeai>=0.8.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
]
